# api/openapi.yaml
openapi: 3.1.0
info:
  title: Aperture API
  version: 0.0.0
  description: |
    REST API for managing Aperture tenants, browser sessions, snapshots, API tokens, and tenant events.

    ## Authentication

    Send `Authorization: Bearer <token>` or authenticate through an account browser session established by OIDC, passkey, or password. Bearer credentials take precedence when both are present. The health check is public. `system:admin` grants every scope. Tenant API tokens may additionally restrict access to explicit session and snapshot IDs. Raw tokens are returned only on creation or rotation.

    ## Tenant selection

    Tenant API tokens use their bound tenant. System administrators and account sessions with multiple memberships select a tenant with `X-Aperture-Tenant-Id` for tenant-scoped operations. Aperture selects the only membership automatically when an account has one.

    ## Pagination

    Lists are newest-first. Pass `meta.nextCursor` as `cursor` while `meta.hasMore` is `true`. Page size defaults to 50 and is capped at 100.

    ## Errors

    Errors use `{ "error": { "code": "...", "message": "..." } }`. Each operation enumerates its possible stable codes; `ErrorCode` is the complete catalog. Request bodies are limited to 1 MiB.
servers:
  - url: /
security:
  - bearerAuth: []
  - webSession: []
...
paths:
...
  /api/sessions:
    parameters:
      - {$ref: '#/components/parameters/SelectedTenantId'}
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List browser sessions
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:read]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [validation_failed, session_service_unavailable]
      description: Newest-first sessions. A resource-restricted token receives only granted sessions; filtering occurs before pagination. Credentials are included only while available and must be treated as secrets.
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/IncludeDeleted'}
        - name: status
          in: query
          description: Exact session lifecycle status.
          schema: {$ref: '#/components/schemas/SessionStatus'}
        - {$ref: '#/components/parameters/TagKey'}
        - {$ref: '#/components/parameters/TagValue'}
        - {$ref: '#/components/parameters/TagOperator'}
      responses:
        '200': {description: Session page, content: {application/json: {schema: {$ref: '#/components/schemas/SessionPage'}}}}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a browser session
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:write]
        conditionalScopes:
          - scope: snapshots:read
            condition:
              source: requestBody
              property: baseSnapshotName
              operator: non_empty
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [invalid_request_body, validation_failed, resource_access_denied, base_snapshot_not_found, base_snapshot_deleted, session_service_unavailable, overlay_mount_failed, browser_start_failed]
      description: Starts a browser with a retained overlay. A resource-restricted token may start a blank session or use a granted base snapshot. The new session is not added to its resource allowlist, but the returned session token remains usable. Failed startup may leave a reopenable `failed` session.
      requestBody: {$ref: '#/components/requestBodies/CreateSession'}
      responses:
        '201': {description: Created session, content: {application/json: {schema: {$ref: '#/components/schemas/CreateSessionResult'}}}}
        default: {$ref: '#/components/responses/Error'}
  /api/sessions/bulk:
    parameters:
      - {$ref: '#/components/parameters/SelectedTenantId'}
    post:
      operationId: getSessionsBulk
      tags: [sessions]
      summary: Get browser sessions by ID
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:read]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [invalid_request_body, validation_failed, session_service_unavailable]
      description: Returns up to 100 sessions in request order. Missing, foreign, deleted, and ungranted sessions are omitted before results are returned.
      requestBody: {$ref: '#/components/requestBodies/SessionBulk'}
      responses:
        '200': {description: Browser sessions, content: {application/json: {schema: {$ref: '#/components/schemas/SessionBulkResponse'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /api/sessions/{sessionId}/session-token/rotate:
    post:
      operationId: rotateSessionToken
      tags: [sessions]
      summary: Rotate a browser session token
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:write]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [validation_failed, resource_access_denied, session_not_found, session_invalid_state, session_expired, session_service_unavailable]
      description: Replaces the access token of a `running` session and invalidates the old token.
      parameters:
        - {$ref: '#/components/parameters/SessionId'}
        - {$ref: '#/components/parameters/SelectedTenantId'}
      responses:
        '200': {description: Session and new token, content: {application/json: {schema: {$ref: '#/components/schemas/SessionMutation'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /api/snapshots:
    parameters:
      - {$ref: '#/components/parameters/SelectedTenantId'}
    get:
      operationId: listSnapshots
      tags: [snapshots]
      summary: List snapshots
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [snapshots:read]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [validation_failed, snapshot_service_unavailable]
      description: Newest-first snapshots. A resource-restricted token receives only granted snapshots; filtering occurs before pagination. Tag predicates use logical AND.
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - name: name
          in: query
          description: ASCII case-insensitive substring matched against the snapshot name.
          schema: {type: string}
          example: signed-in
        - {$ref: '#/components/parameters/Deleted'}
        - {$ref: '#/components/parameters/LegacyIncludeDeleted'}
        - {$ref: '#/components/parameters/TagKey'}
        - {$ref: '#/components/parameters/TagValue'}
        - {$ref: '#/components/parameters/TagOperator'}
      responses:
        '200': {description: Snapshot page, content: {application/json: {schema: {$ref: '#/components/schemas/SnapshotPage'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /api/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List tenant events
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:read]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [validation_failed, event_service_unavailable]
      description: Newest-first tenant events. A resource-restricted token receives only events for granted sessions and snapshots; filtering occurs before pagination. `resourceType` and `resourceId` are exact-match filters.
      parameters:
        - {$ref: '#/components/parameters/SelectedTenantId'}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - name: resourceType
          in: query
          description: Exact resource category, such as `session`.
          schema: {type: string}
          example: session
        - name: resourceId
          in: query
          description: Exact resource identifier within the selected tenant.
          schema: {type: string}
          example: 019f6cf0-0000-7000-8000-000000000010
      responses:
        '200': {description: Event page, content: {application/json: {schema: {$ref: '#/components/schemas/EventPage'}}}}
        default: {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Aperture API token
      description: |
        Aperture API token sent as `Authorization: Bearer <token>`. API tokens normally start with `apt_`. Session access tokens use a separate `aps_` format and are not accepted as API authentication.
    webSession:
      type: apiKey
      in: cookie
      name: aperture_session
      description: Server-side browser session established by OIDC, passkey, or password login.
  parameters:
    Limit:
      name: limit
      in: query
      description: Requested page size. Omitted or `0` uses 50. Values above 100 are accepted but clamped to 100.
      schema:
        type: integer
        minimum: 0
        default: 50
      example: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque `nextCursor` from the previous page. Reuse it only with the same operation and filters.
      schema: {type: string}
      example: eyJjcmVhdGVkQXQiOiIyMDI2LTA3LTE3VDEyOjAwOjAwWiIsImlkIjoiMDE5ZjZjZjAtMDAwMC03MDAwLTgwMDAtMDAwMDAwMDAwMDAwIn0
    IncludeDeleted:
      name: includeDeleted
      in: query
      description: Include soft-deleted sessions when `true`. The default is `false`.
      schema: {type: boolean, default: false}
    LegacyIncludeDeleted:
      name: includeDeleted
      in: query
      deprecated: true
      description: Legacy switch that includes active and deleted resources when `true`. Ignored when `deleted` is supplied; prefer `deleted=all`.
      schema: {type: boolean, default: false}
    Deleted:
      name: deleted
      in: query
      description: Select active resources, deleted resources only, or both. The default is `active`.
      schema:
        type: string
        enum: [active, deleted, all]
        default: active
...
    SelectedTenantId:
      name: X-Aperture-Tenant-Id
      in: header
      description: |
        Tenant selected for a tenant-scoped operation. System administrators and account sessions may provide this header. A tenant API token uses its bound tenant and may omit the header; selecting a different tenant is forbidden.
      schema: {$ref: '#/components/schemas/UUIDv7'}
      example: 019f6cf0-0000-7000-8000-000000000001
...
    SessionId:
      name: sessionId
      in: path
      required: true
      description: UUIDv7 identifier of the browser session.
      schema: {$ref: '#/components/schemas/UUIDv7'}
...
    TagKey:
      name: tagKey
      in: query
      description: |
        Repeatable tag key. Each entry pairs by position with `tagValue` and, when supplied, `tagOperator`. Multiple predicates are combined with logical AND.
      schema: {type: array, items: {type: string, minLength: 1}}
      style: form
      explode: true
      example: [environment, region]
    TagValue:
      name: tagValue
      in: query
      description: |
        Repeatable tag value paired with `tagKey`. `eq` and `neq` accept one value. `in` and `not_in` accept comma-separated values in the matching entry.
      schema: {type: array, items: {type: string, minLength: 1}}
      style: form
      explode: true
      example: [production, 'eu-west-1,eu-central-1']
    TagOperator:
      name: tagOperator
      in: query
      description: |
        Optional operator paired by position with each tag predicate. Omit the entire parameter to use `eq` for every predicate. If supplied, its entry count must match `tagKey`.
      schema:
        type: array
        items:
          type: string
          enum: [eq, neq, in, not_in]
      style: form
      explode: true
      example: [eq, in]
  requestBodies:
...
    CreateSession:
      required: true
      description: Browser session configuration. The JSON body must not exceed 1 MiB.
      content: {application/json: {schema: {$ref: '#/components/schemas/CreateSessionInput'}}}
    SessionBulk:
      required: true
      description: Ordered UUIDv7 session IDs to resolve. The JSON body must not exceed 1 MiB.
      content: {application/json: {schema: {$ref: '#/components/schemas/SessionBulkInput'}}}
...
  responses:
    Error:
      description: API error. Branch on the HTTP status and stable `error.code`.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            invalidRequestBody:
              summary: 400 · invalid JSON body
              value: {error: {code: invalid_request_body, message: invalid request body}}
            validationFailed:
              summary: 400 · invalid field or parameter
              value: {error: {code: validation_failed, message: 'validation error\nname is required'}}
            tenantSelectionRequired:
              summary: 400 · tenant header required
              value: {error: {code: tenant_selection_required, message: tenant selection required}}
            authenticationRequired:
              summary: 401 · bearer token missing
              value: {error: {code: authentication_required, message: authentication required}}
            authenticationTokenExpired:
              summary: 401 · bearer token expired
              value: {error: {code: authentication_token_expired, message: authentication token expired}}
            insufficientScope:
              summary: 403 · required scope missing
              value: {error: {code: insufficient_scope, message: insufficient scope}}
            resourceAccessDenied:
              summary: 403 · resource is not granted
              value: {error: {code: resource_access_denied, message: resource access denied}}
            tenantSelectionNotPermitted:
              summary: 403 · tenant token selected another tenant
              value: {error: {code: tenant_selection_not_permitted, message: tenant selection not permitted}}
            sessionNotFound:
              summary: 404 · session missing or foreign
              value: {error: {code: session_not_found, message: session not found}}
            sessionInvalidState:
              summary: 409 · operation invalid for current state
              value: {error: {code: session_invalid_state, message: session is in an invalid state for this operation}}
            sessionExpired:
              summary: 410 · retained session expired
              value: {error: {code: session_expired, message: session expired}}
            internalError:
              summary: 500 · unexpected server failure
              value: {error: {code: internal_error, message: internal server error}}
            browserStartFailed:
              summary: 502 · browser failed to start
              value: {error: {code: browser_start_failed, message: browser failed to start}}
  schemas:
    UUIDv7:
      type: string
      format: uuid
      x-uuid-version: 7
      description: Time-ordered UUID version 7 generated by Aperture.
      example: 019f6cf0-0000-7000-8000-000000000001
...
    SessionStatus:
      type: string
      description: |
        Session lifecycle state:

        - `creating`: browser startup is in progress.
        - `running`: browser and routing are available.
        - `suspended`: browser is intentionally stopped while retained state remains available.
        - `deleted`: session is soft-deleted and may be reopened while retained.
        - `expired`: retention elapsed and the session can no longer be used.
        - `failed`: startup or reopen failed; retained state may still be reopenable.
      enum: [creating, running, suspended, deleted, expired, failed]
      example: running
    StringMap:
      type: object
      description: Arbitrary string key-value metadata. Tag-writing operations additionally reject blank keys and blank values.
      additionalProperties: {type: string}
      example: {environment: production, region: eu-west-1}
    PageMeta:
      type: object
      description: Cursor pagination metadata for a newest-first result page.
      required: [limit, hasMore]
      properties:
        limit:
          type: integer
          description: Effective page size after applying the default and maximum.
          minimum: 1
          maximum: 100
          example: 50
        nextCursor:
          type: string
          description: Opaque cursor for the next page. Omitted when there is no next page.
          example: eyJjcmVhdGVkQXQiOiIyMDI2LTA3LTE3VDEyOjAwOjAwWiIsImlkIjoiMDE5ZjZjZjAtMDAwMC03MDAwLTgwMDAtMDAwMDAwMDAwMDAwIn0
        hasMore:
          type: boolean
          description: Whether another page is available.
          example: true
...
    Error:
      type: object
      description: Standard error envelope returned by REST API operations.
      required: [error]
      properties: {error: {$ref: '#/components/schemas/ErrorDetail'}}
...
    CreateSessionInput:
      type: object
      description: Configuration for a new retained browser session.
      required: [browser]
      properties:
        baseSnapshotName:
          type: [string, 'null']
          description: Active snapshot used as the browser filesystem base. Omit, send `null`, or send an empty value for a blank session.
          example: signed-in-base
        label:
          type: [string, 'null']
          description: Optional operator-facing label for the session.
          example: checkout smoke test
        browser: {$ref: '#/components/schemas/BrowserConfig'}
        tags:
          allOf: [{$ref: '#/components/schemas/StringMap'}]
          description: Initial session tags.
      example:
        baseSnapshotName: signed-in-base
        label: checkout smoke test
        browser: {channel: chromium, mode: headless, args: ['--lang=en-US']}
        tags: {environment: staging, suite: checkout}
...
    Session:
      type: object
      description: Browser session state, retention metadata, tags, and currently available access credentials.
      required: [id, tenantId, status, browser, capabilities, connection, createdAt, expiresAt]
      properties:
        id:
          allOf: [{$ref: '#/components/schemas/UUIDv7'}]
          description: Session identifier.
        tenantId:
          allOf: [{$ref: '#/components/schemas/UUIDv7'}]
          description: Tenant that owns the session.
        baseSnapshotName:
          type: string
          description: Snapshot name used as the session base. Omitted for a blank session.
          example: signed-in-base
        label:
          type: string
          description: Optional operator-facing session label.
          example: checkout smoke test
        status: {$ref: '#/components/schemas/SessionStatus'}
        browser: {$ref: '#/components/schemas/BrowserConfig'}
        capabilities: {$ref: '#/components/schemas/SessionCapabilities'}
        connection: {$ref: '#/components/schemas/SessionConnection'}
        createdAt:
          type: string
          format: date-time
          description: Time the session record was created.
          example: '2026-07-17T12:00:00Z'
        startedAt:
          type: string
          format: date-time
          description: Most recent successful browser start time. Omitted before startup succeeds.
          example: '2026-07-17T12:00:03Z'
        stoppedAt:
          type: string
          format: date-time
          description: Most recent browser stop time. Omitted while no stop has occurred.
          example: '2026-07-17T13:00:00Z'
        deletedAt:
          type: [string, 'null']
          format: date-time
          description: Soft-deletion time, or `null` for a session that has not been deleted.
          example: null
        expiresAt:
          type: string
          format: date-time
          description: Retention deadline after which the session and its overlay may be purged. This is not token expiry.
          example: '2026-07-24T12:00:00Z'
        lastConnectedAt:
          type: string
          format: date-time
          description: Most recent routed client activity time. Omitted until a client connects.
          example: '2026-07-17T12:30:00Z'
        suspendedAt:
          type: string
          format: date-time
          description: Most recent suspension time. Omitted unless the session is suspended.
          example: '2026-07-17T13:00:00Z'
        tags:
          allOf: [{$ref: '#/components/schemas/StringMap'}]
          description: Session metadata tags. Omitted when no tags are set.
        collaboration:
          $ref: '#/components/schemas/SessionCollaborationCapabilities'
...
    SessionPage:
      type: object
      description: Cursor-paginated browser sessions.
      required: [data, meta]
      properties: {data: {type: array, items: {$ref: '#/components/schemas/Session'}}, meta: {$ref: '#/components/schemas/PageMeta'}}
    SessionBulkInput:
      type: object
      description: Ordered session IDs for a tenant-scoped bulk lookup. Missing, foreign, and deleted IDs are omitted from the result.
      required: [ids]
      properties:
        ids:
          type: array
          description: Up to 100 unique UUIDv7 values. An empty array is valid.
          maxItems: 100
          uniqueItems: true
          items: {$ref: '#/components/schemas/UUIDv7'}
          example: ['019f6cf0-0000-7000-8000-000000000010', '019f6cf0-0000-7000-8000-000000000011']
    SessionBulkResponse:
      type: object
      description: Sessions found for the requested IDs, preserving request order.
      required: [sessions]
      properties:
        sessions: {type: array, items: {$ref: '#/components/schemas/Session'}}
    CreateSessionResult:
      type: object
      description: Newly created session. Active connection data is nested in the session.
      required: [session]
      properties:
        session: {$ref: '#/components/schemas/Session'}
    SessionMutation:
      type: object
      description: Updated session. Active connection data is nested in the session.
      required: [session]
      properties:
        session: {$ref: '#/components/schemas/Session'}
...
    SnapshotPage:
      type: object
      description: Cursor-paginated snapshots.
      required: [data, meta]
      properties: {data: {type: array, items: {$ref: '#/components/schemas/Snapshot'}}, meta: {$ref: '#/components/schemas/PageMeta'}}
...
    EventPage:
      type: object
      description: Cursor-paginated tenant events.
      required: [data, meta]
      properties: {data: {type: array, items: {$ref: '#/components/schemas/Event'}}, meta: {$ref: '#/components/schemas/PageMeta'}}